fix: ConvoStarter UI
